train first variational autoencoder (VAE)
classification branch is incorret, because it takes z (a random distribution as input). It makes the classification process harder and causes the assertion error present in notebook.

Also, in VAE tutorials I found, only the mean (z_mean) is used to visualize the latent space. I'll try that in the next commits.